+++ src/sound.ts
@@ -2,8 +2,9 @@ import correctSoundUrl from './audio/correct.mp3';
 import wrongSoundUrl from './audio/wrong.mp3';
 import gameOverSoundUrl from './audio/game-over.mp3';
 
-import music1Url from './audio/bttf-1.mp3';
-import music2Url from './audio/bttf-2.mp3';
+import f1Url from './audio/f1.mp3';
+import bttfUrl1 from './audio/bttf-1.mp3';
+import bttfUrl2 from './audio/bttf-2.mp3';
 
 const soundBlobs: SoundBlobs = {
   correct: null,
@@ -13,7 +14,10 @@ const soundBlobs: SoundBlobs = {
 
 let musicIndex = 0;
 let musicAudio: HTMLAudioElement | null = null;
-const musicBlobs: Array<Blob> = [];
+const musicBlobs: Record<AnimationType, Array<Blob>> = {
+  horse: [],
+  formulaCar: []
+};
 
 export async function initSounds(): Promise<void> {
   if (!Audio) {
@@ -30,10 +34,8 @@ export async function initMusic(): Promise<void> {
   if (!Audio) {
     return;
   }
-  musicBlobs.push(
-    await loadAudioBlob(music1Url),
-    await loadAudioBlob(music2Url)
-  );
+  musicBlobs.horse.push(await loadAudioBlob(bttfUrl1), await loadAudioBlob(bttfUrl2));
+  musicBlobs.formulaCar.push(await loadAudioBlob(f1Url));
 }
 
 async function loadAudioBlob(url: string): Promise<Blob> {
@@ -55,11 +57,12 @@ export async function playSound(sound: Sound): Promise<void> {
   await playAudio(audio);
 }
 
-export async function playMusic(singleTrack = false): Promise<void> {
+export async function playMusic(animationType: AnimationType, singleTrack = false): Promise<void> {
   if (musicAudio) {
     return;
   }
-  const audioBlob = musicBlobs[musicIndex];
+  const musicItems = musicBlobs[animationType];
+  const audioBlob = musicItems[musicIndex];
   if (!audioBlob) {
     return;
   }
@@ -71,10 +74,10 @@ export async function playMusic(singleTrack = false): Promise<void> {
     return;
   }
   musicIndex++;
-  if (musicIndex >= musicBlobs.length) {
+  if (musicIndex >= musicItems.length) {
     musicIndex = 0;
   }
-  playMusic().catch(e => console.error(e));
+  playMusic(animationType).catch(e => console.error(e));
 }
 
 export async function stopMusic(): Promise<void> {