+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
+- Different animation themes are supported — Formula 1 theme is introduced.
 - Logic error in `createDivisionOperation` is fixed.
 - Sprite opacity flickering on fade out is fixed.
+- Maximum challenges count is increased up to 7.
 - Dependencies are upgraded.
 - Remote repository is migrated to https://github.com/ezze/math-hunger.